I'm going to revamp the book page + fix ratings at the same time.
